# target/classes/swagger.json
{"swagger":"2.0","info":{"description":"Telephone Directory API reference for developers","version":"2018.12","title":"Telephone Directory using Spring Boot","termsOfService":"https://github.com/chandanv89/Telephone-Directory-Spring-Boot/blob/master/LICENSE","contact":{"name":"Chandan Veerabhadrappa","url":"https://github.com/chandanv89"},"license":{"name":"MIT License","url":"https://github.com/chandanv89/Telephone-Directory-Spring-Boot/blob/master/LICENSE"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"directory","description":"Contact Numbers Controller"},{"name":"telephone-directory","description":"Contact Numbers Controller"}],"paths":{"/contactnumbers":{"get":{"tags":["directory","telephone-directory"],"summary":"getAllContactNumbers","description":"Fetch all the ContactNumbers in the directory.","operationId":"getAllContactNumbersUsingGET_7","produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ApiResponse"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/contacts":{"post":{"tags":["directory","telephone-directory"],"summary":"addContacts","operationId":"addContactsUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"contacts","description":"contacts","required":true,"schema":{"type":"array","items":{"$ref":"#/definitions/Contact"}}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ApiResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/contacts/":{"get":{"tags":["directory","telephone-directory"],"summary":"getAllContacts","description":"fetches all the contacts from the directory","operationId":"getAllContactsUsingGET","produces":["application/json","application/jasoon"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ApiResponse"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/emails":{"get":{"tags":["directory","telephone-directory"],"summary":"/getAllEmails","operationId":"getAllEmailsUsingGET_7","produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ApiResponse"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}}},"definitions":{"ApiResponse":{"type":"object","properties":{"body":{"type":"object"},"status":{"type":"string","enum":["100","101","102","103","200","201","202","203","204","205","206","207","208","226","300","301","302","303","304","305","307","308","400","401","402","403","404","405","406","407","408","409","410","411","412","413","414","415","416","417","418","419","420","421","422","423","424","426","428","429","431","451","500","501","502","503","504","505","506","507","508","509","510","511"]}},"title":"ApiResponse"},"Contact":{"type":"object","properties":{"contactNumbers":{"type":"array","items":{"$ref":"#/definitions/ContactNumber"}},"createdOn":{"type":"string"},"emails":{"type":"array","items":{"$ref":"#/definitions/Email"}},"firstName":{"type":"string"},"fullName":{"type":"string"},"id":{"type":"string"},"isDeleted":{"type":"string"},"lastName":{"type":"string"},"updatedOn":{"type":"string"}},"title":"Contact"},"ContactNumber":{"type":"object","properties":{"category":{"type":"string"},"createdOn":{"type":"string"},"id":{"type":"string"},"number":{"type":"string"},"parentContactId":{"type":"string"},"primary":{"type":"boolean"},"updatedOn":{"type":"string"}},"title":"ContactNumber"},"Email":{"type":"object","properties":{"category":{"type":"string"},"createdOn":{"type":"string"},"emailId":{"type":"string"},"id":{"type":"string"},"parentContactId":{"type":"string"},"primary":{"type":"boolean"},"updatedOn":{"type":"string"}},"title":"Email"}}}